Add files via upload
Given a partially filled 9×9 2D array ‘grid[9][9]’, the goal is to assign digits (from 1 to 9) to the empty cells so that every row, column, and sub grid of size 3×3 contains exactly one instance of the digits from 1 to 9. similarly it can be applied to a 4×4 grid ( sub grid of size 2×2 ).